+++ library/Episciences/Mail/RemindersManager.php
@@ -297,7 +297,8 @@ class Episciences_Mail_RemindersManager
             'name' => 'delay',
             'label' => $tooltip . $translator->translate('Délai'),
             'required' => true,
-            'value' => ($reminder) ? $reminder->getDelay() : null
+            'value' => ($reminder) ? $reminder->getDelay() : null,
+            'description' => $tooltipMsg
         ]));
 
         // Select: Répétition ******************************************************